requirements.txt
apache-airflow==2.3.4
# The amazon provider can be found in the Docker base image. Alternatively, run
# docker exec -i ANY_AIRFLOW_CONTAINER pip list | grep amazon
# against any of the Airflow containers.
apache-airflow-providers-amazon[cncf.kubernetes]==5.1.0
apache-airflow-providers-docker==3.12.2 
